function to add new products to firebase

+++ app/src/main/java/com/example/brewco/data/model/Product.kt
@@ -6,5 +6,5 @@ data class Product(
     val categoria: String = "",
     val inventario: Int = 0,
     val inventario_minimo: Int = 0,
-    val precio: Double = 0.0
+    val precio: Int = 0
 )
